@@ -1,47 +1,138 @@
 <template>
-  <Hello :msg="`Electron ${ state.version }`" />
-  <Hello msg="Vue 3.0 + Vite" />
-  <div>
-    <button style="width: 60%;" @click="increment">Incrementa</button>
-  </div>
-  <p>count: {{state.count}}</p>
-  <p>double: {{state.double}}</p>
+  <div class="flex flex-col mx-auto p-4">
+      <div class="flex justify-between">
+        <h1 class="h1 text-2xl text-gray-800 tracking-wider uppercase">WSL2 Linux Distributions </h1>
+        <span @click="checkWSL()" class="flex items-center justify-center text-lg text-gray-400 hover:text-green-500">
+          <svg xmlns="http://www.w3.org/2000/svg"
+               fill="none"
+               viewBox="0 0 24 24"
+               stroke="currentColor"
+               stroke-linecap="round"
+               stroke-linejoin="round"
+               stroke-width="2"
+               class="h-6 w-6">
+            <path d="M4 4v5h.582m15.356 2A8.001 8.001 0 004.582 9m0 0H9m11 11v-5h-.581m0 0a8.003 8.003 0 01-15.357-2m15.357 2H15" />
+          </svg>
+        </span>
+      </div>
+      <div class="flex flex-col mt-2">
+        <div class="-my-2 overflow-x-auto sm:-mx-6 lg:-mx-8">
+          <div class="py-2 align-middle inline-block min-w-full sm:px-6 lg:px-8">
+            <div class="shadow overflow-hidden border-b border-gray-200 sm:rounded-lg">
+              <table class="min-w-full divide-y divide-gray-200">
+                <thead>
+                <tr>
+                  <th scope="col" class="px-6 py-3 bg-gray-50 text-left text-xs font-medium text-gray-500 uppercase tracking-wider">
+                    Name
+                  </th>
+                  <th scope="col" class="px-6 py-3 bg-gray-50 text-left text-xs font-medium text-gray-500 uppercase tracking-wider">
+                    Version
+                  </th>
+                  <th scope="col" class="px-6 py-3 bg-gray-50 text-left text-xs font-medium text-gray-500 uppercase tracking-wider">
+                    Status
+                  </th>
+                  <th scope="col" class="px-6 py-3 bg-gray-50 text-left text-xs font-medium text-gray-500 uppercase tracking-wider">
+                    Actions
+                  </th>
+                </tr>
+                </thead>
+                <tbody class="bg-white divide-y divide-gray-200">
+                <tr v-for="distribution in state.distributions" :key="distribution.name">
+                  <td class="px-6 py-4 whitespace-nowrap">
+                    <div class="text-sm font-medium text-gray-900">
+                      {{ distribution.name }}
+                    </div>
+                  </td>
+                  <td class="px-6 py-4 whitespace-nowrap">
+                    <div class="text-sm text-gray-900">{{ distribution.version }}</div>
+                  </td>
+                  <td class="px-6 py-4 whitespace-nowrap">
+                    <span class="px-2 inline-flex text-xs leading-5 font-semibold rounded-full" :class="distribution.status === 'Running' ? 'bg-green-100 text-green-800' : 'bg-red-100 text-red-800'">
+                     {{ distribution.status }}
+                    </span>
+                  </td>
+                  <td class="px-6 py-4 whitespace-nowrap text-sm font-medium flex">
+                    <span class="flex items-center justify-center text-lg text-gray-400 hover:text-green-500">
+                        <svg xmlns="http://www.w3.org/2000/svg"
+                             fill="none"
+                             viewBox="0 0 24 24"
+                             stroke="currentColor"
+                             stroke-linecap="round"
+                             stroke-linejoin="round"
+                             stroke-width="2"
+                             class="h-6 w-6">
+                          <path d="M14.752 11.168l-3.197-2.132A1 1 0 0010 9.87v4.263a1 1 0 001.555.832l3.197-2.132a1 1 0 000-1.664z" />
+                          <path d="M21 12a9 9 0 11-18 0 9 9 0 0118 0z" />
+                        </svg>
+                      </span>
+                    <span class="flex items-center justify-center text-lg text-gray-400 hover:text-yellow-500">
+                      <svg xmlns="http://www.w3.org/2000/svg"
+                           fill="none"
+                           viewBox="0 0 24 24"
+                           stroke="currentColor"
+                           stroke-linecap="round"
+                           stroke-linejoin="round"
+                           stroke-width="2"
+                           class="h-6 w-6">
+                        <path d="M21 12a9 9 0 11-18 0 9 9 0 0118 0z" />
+                        <path d="M9 10a1 1 0 011-1h4a1 1 0 011 1v4a1 1 0 01-1 1h-4a1 1 0 01-1-1v-4z" />
+                      </svg>
+                    </span>
+                    <span class="flex items-center justify-center text-lg text-gray-400 hover:text-red-500">
+                      <svg xmlns="http://www.w3.org/2000/svg"
+                           fill="none"
+                           viewBox="0 0 24 24"
+                           stroke="currentColor"
+                           stroke-linecap="round"
+                           stroke-linejoin="round"
+                           stroke-width="2"
+                           class="h-6 w-6">
+                        <path d="M19 7l-.867 12.142A2 2 0 0116.138 21H7.862a2 2 0 01-1.995-1.858L5 7m5 4v6m4-6v6m1-10V4a1 1 0 00-1-1h-4a1 1 0 00-1 1v3M4 7h16" />
+                      </svg>
 
-  <div class="bg-white rounded-lg p-6">
-       <img class="h-16 w-16 rounded-full mx-auto">
-    <div class="text-center">
-         <h2 class="text-lg">Erin Lindford</h2>
-         <div class="text-purple-500">Product Engineer</div>
-         <div class="text-gray-600">erinlindford@example.com</div>
-         <div class="text-gray-600">(555) 765-4321</div>
-       </div>
-     </div>
+                    </span>
+                  </td>
+                </tr>
+                </tbody>
+              </table>
+            </div>
+          </div>
+        </div>
+      </div>
+    </div>
 </template>
 
 <script lang='ts'>
 import Hello from '../components/HelloWorld.vue'
-import { reactive, computed } from 'vue'
-const version = require('process').versions.electron
+import Toggle from '../components/Toggle.vue'
+import Tooltip from '../components/Tooltip.vue'
+import { reactive, onMounted } from 'vue'
+import { sendSync } from '../store/ipc'
 
 export default {
-  name: 'Index',
+  name: 'Home',
   components: {
-    Hello
+    Hello,
+    Toggle,
+    Tooltip
   },
   setup() {
     const state:any = reactive({
-      count: 0,
-      double: computed(() => state.count * 2),
-      version: version
+      distributions: null,
+      showToolTip: false
     })
 
-    function increment() {
-      state.count++
+    onMounted(() => {
+      checkWSL()
+    })
+
+    function checkWSL() {
+      state.distributions =  sendSync('checkWSL').splice(1)
     }
 
     return {
       state,
-      increment,
+      checkWSL
     }
   },
 }
